tsb: Add basic Toshiba processor bring up code
This is a first pass at a basic port for NuttX on the APBridge. It
includes preliminary configuration and processor bringup code for APB1
on the Quanta Big Dev Board. No board-specific initialization is done
and most options for NuttX are disabled.

Boots to usermode hello world main() test code.

The binary is linked to run out of RAM, located at 0x10000000 on the
APBridge.

The system controller and UART bringup code is copied from Toshiba
reference firmware. I plan to clean this up and abstract it out in a
future changeset.

-Set up scheduler for 100ms ticks
-No peripheral interrupts yet

This changeset applies to upstream NuttX tag 'nuttx-7.5'

Testing Done: boot hello world program on APB1 on BDB
Bug Number: SW-193
Review URL: https://reviews.leaflabs.com/r/145/
Reviewed by: fparent, abailon