mm: make memory allocation atomic
The synopsys USB driver is requiring to be able to do atomic
memory allocation. This commit takes advantage of the fact that
the TSB/SVC processors are uniprocessors, it cannot work on a multiprocessors.

Testing Done: Tested the USB driver with this patch and it seems to work fine
Bug Number: SW-197
Review URL: https://reviews.leaflabs.com/r/146/
Reviewed by: perry