serial: uart_16550.c: Fix LCR bits settings
The LCR bits setting needs to be changed depending of the user
configuration.
Unfortunately, it is always set to 7.
Adapt the LCR bits to the input value number of bits.

Testing Done: Output correctly the characters on a 8 bits console
Bug Number: SW-201
Review URL: https://reviews.leaflabs.com/r/152/
Reviewed by: fparent

Signed-off-by: Benoit Cousson <bcousson@baylibre.com>