bdb: apb: defconfig: Enable and configure 16550_UART
Enable the generic 16550 driver and provide all the details needed
for the driver to work properly on the AP/GP bridges.

The defconfig must then be used to recreate the .config with all
the needed settings.

Testing Done: Use this UART as the NSH console
Bug Number: SW-200
Review URL: https://reviews.leaflabs.com/r/153/
Reviewed by: perry

Signed-off-by: Benoit Cousson <bcousson@baylibre.com>