tools: configure.sh: fix CONFIG_APPS_DIR defconfig detection bug
configure.sh should not add an APPS_DIR section if one is included in
the defconfig.

This fixes a bug where it erroneously adds CONFIG_APPS_DIR to the
generated .config even when one is provided in the defconfig.

Testing Done: configure project for apb1, see that it does not add
    junk
Bug Number: SW-242
Review URL: https://reviews.leaflabs.com/r/181/
Reviewed by: bcousson