tsb: unipro: unipro controller driver core
This is a first pass at a unipro stack for the Toshiba bridge chips.

The bridge chips are operated in mode 1, a bizarre mode where the sender
gets to pick the destination address of the packet. This will allow us
to implement flow control which the hardware does not provide us.

Users can register their driver with the core on a per-cport basis.
Registered drivers can send data through a cport, or be notified when a
packet is received on a cport. These callbacks are received in irq
context.

This patch does not implement flow control, so be careful not to overrun
your receiver.

Testing Done: send unipro packets back and forth, check that they arrive
    correctly and handlers are called
Bug Number: SW-106
Review URL: https://reviews.leaflabs.com/r/190/
Reviewed by: mgreer