BDB: SVC: AP <-> LCD demo code, from Quanta's reference code
Setup the routes and the connection for the AP to LCD demo.
The code is inspired from the Quanta demo code; the routes and
default connection parameters are hard coded, some delays are used
to allow the switch and the bridges to settle.

Testing Done: Check routing table and bridges ports
Bug Number: SW-74
Review URL: https://reviews.leaflabs.com/r/226/
Reviewed by: perry

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>