SVC: debug: make dbg_pr() expansion safer
The dbg_pr() macro contains an if statement with no else.  If it is
called within another if statement with its own else, the else won't
parse the way it's intended.

Enclose the expansion of dbg_pr() to avoid this problem.

Testing Done: check console output on BDB1A
Bug Number: -
Review URL:
Reviewed by:

Signed-off-by: Alex Elder <elder@linaro.org>
Acked-by: Jean Pihet <jean.pihet@newoldbits.com>