SVC: Endo: Update to the new interface API for power control
This allows the use of the interface abstraction layer by the power control
app.

Note: the interface support for the Endo is temporary and is to be replaced by
the proper implementation of the interface API via the board file.

Testing Done: test power supplies and wakeout signals on BDB2A and BDB1B
Bug Number: SW-183
Review URL: https://reviews.leaflabs.com/r/337/
Reviewed by: fab, perry, mag

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>