svc: tsb_switch: truly insane tsb_switch_es1.c log output.
This is needed to share detailed diagnostics with Toshiba about some
BDB1Bs that have problems transitioning to HS-G1 on some links. Output
is reduced to reasonable levels when DBG_INSANE is not used.

Testing Done: build and boot to NuttShell on board that is not
              behaving properly; verify output.
Bug Number: SW-486
Review URL: https://reviews.leaflabs.com/r/346/
Reviewed by: perry, fparent

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>