svc: tsb_switch: use new link power config at startup
Use the HS and PWM gear helpers in switch_link_power_set_default()
instead of hard-coding HS-G1. Move the new default configuration to
the top of the file.

To confirm that this preserves the old behavior at HS-G1, here is a
diff of the old and new behavior obtained with a DBG_VERBOSE setting
in dbg_set_config(). Here are the details with an explanation of why I
think the few differences that do occur are correct or improvements.

    --- /tmp/before.txt	2015-02-23 23:42:08.463024638 -0500
    +++ /tmp/after.txt	2015-02-23 23:43:05.599025633 -0500
    @@ -11,17 +11,18 @@
     [DBG_SWITCH] es1_peer_set(): portid=1, (attr,sel)=(0x4022,5)=0x0000, ret=0x00
     [DBG_SWITCH] es1_peer_set(): portid=1, (attr,sel)=(0x4021,5)=0x0001, ret=0x00
     [DBG_SWITCH] es1_peer_set(): portid=1, (attr,sel)=(0x4020,5)=0x0001, ret=0x00
    +[DBG_SWITCH] switch_configure_link(): port=0

Harmless extra log message.

     [DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x1568,0)=0x0001, ret=0x00
     [DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x1569,0)=0x0001, ret=0x00
    -[DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x156a,0)=0x0001, ret=0x00

The new version deletes an explicit set to PA_HSSeries (0x156a) to its
reset value of series A = 1.

There are lot of restrictions in setting PA_HSSeries that we do not
enforce, so the new version just makes sure the caller leaves this
attribute at its reset value. We can handle rate group switching
properly when we are ready, as this is just a nice-to-have (rate B
series are only about 17% higher than the corresponding rate A
series).

     [DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x1560,0)=0x0002, ret=0x00
     [DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x1583,0)=0x0001, ret=0x00
     [DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x1584,0)=0x0001, ret=0x00
     [DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x1580,0)=0x0002, ret=0x00
    +[DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x1585,0)=0x0000, ret=0x00

The new routine sets PA_Scrambling (0x1585) to its reset value of
OFF=0. The UniPro spec says a scrambling request is made only if this
is explicitly set to 1. So the behavior for the new version should be
the same.

     [DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x15b0,0)=0x1fff, ret=0x00
     [DBG_SWITCH] es1_set(): portid=0, (attr,sel)=(0x1571,0)=0x0011, ret=0x00
     [DBG_SWITCH] es1_get(): portid=0, (attr,sel)=(0xd040,0)=0x0002, ret=0x00
    -[DBG_SWITCH] es1_peer_set(): portid=0, (attr,sel)=(0xd089,0)=0x0118, ret=0x08
    +[DBG_SWITCH] es1_peer_get(): portid=0, (attr,sel)=(0x4020,0)=0x0000, ret=0x08

The old link configuration set TSB_MAXSEGMENTCONFIG (0xd089) to 0x118,
naively assuming that the link power configuration succeeded merely
because a success code was retrieved from TSB_DME_POWERMODEIND ;).

The new version is older and wiser, and attempts to check
T_CONNECTIONSTATE (0x4020) for CPort 0 before it really trusts that
the link was reconfigured.

In either case, the attribute write fails with result code 8 (as
expected; I'm using this to debug a bad board).

     [DBG_SWITCH] es1_peer_set(): portid=1, (attr,sel)=(0x4020,5)=0x0000, ret=0x00
     [DBG_SWITCH] es1_peer_set(): portid=0, (attr,sel)=(0x4020,0)=0x0000, ret=0x08
     [DBG_SWITCH] switch_connection_create: Connection setup failed. [0:0]<->[1:5] TC: 0 Flags: 6 rc: 8

Testing Done: see above
Bug Number: SW-486
Review URL: https://reviews.leaflabs.com/r/349/
Reviewed by: perry

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>