apps: slice: Add example manifest which can be read from host
Proof of concept that allows the I2C host read the raw greybus
handshake from slice_cmd. For the purposes of this example, a
virtual "SVC message register" is created and pre-populated with
the greybus handshake. For now, writes to this register are
ignored.

Please Note:
In order to compile slice_cmd, developers will now need to have
manifesto compiler on their PATH. See this page for details:
https://github.com/projectara/Firmware-wiki/wiki/Development-Machine-Setup#manifesto

Signed-off-by: Greg Meiste <w30289@motorola.com>
Change-Id: I732a42c7b49687df333fe026d7c8c070b1bc98f7