apps: svc: add "linktest" sub-command
WARNING: this is not thread-safe. Run with switch IRQs disabled only.

Extend the "svc" app with commands for testing either an individual
port, or all of the switch ports with built-in peer UniPro on the
switch.

Testing Done: Tested on BDB2A
Bug Number: SW-487
Review URL: https://reviews.leaflabs.com/r/552/
Reviewed by: Jean Pihet, Perry Hung

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>