tsb: Add Device Core support
Add simple device core infrastructure for low-level
drivers.  It provides a distinct interface between the
low-level drivers and their users.  Currently, only
one user may have a device open at a time.

A device driver registers with the device core by passing
a device driver structure to device_register_driver().
The device driver structure contains a class and a name
member which specify the class of hardware and the specific
type of hardware within the class that it supports.  It also
specifies the generic driver interface routines (i.e., probe,
remove, open, and close) and the class-specific routines.

Actual devices and their resources are placed in a table and
the table is registered using the device_table_register() routine.
When a device driver is registered, the table is searched for
devices that the driver supports.  The driver's probe routine
will be called for each matching device.

When a user wants to access a device, it calls device_open()
with the class and id (the instance number for devices in that
class).  When finished, it calls device_close() with the device
structure that was returned by the corresponding device_open().

Testing Done: Tested during I2S testing
Bug Number: SW-337
Review URL:
Reviewed by:

Signed-off-by: Mark A. Greer <mgreer@animalcreek.com>