greybus: core: use auto var for timedout responses
In the hypothesis we are running low on available memory on the heap, use
a response allocated on the stack. Being able to call the callback of an
operation correctly is important and is not an operation that can be dropped,
even when the system if out of free memory.

Testing Done: none
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/392/
Reviewed by: mgreer, mbolivar