ara/lgd: start AP without USB
USB was required to power on the AP.
The APbridge is responsible when USB is unplugged to power on the AP using ONKEY gpio.
That was done but the time between ONKEY set and ONKEY reset was to short.
In addition, the gpio controller was not enable (enabled later by Greybus).
Initialize gpio controller and add a delay to solve the issue.

Testing Done: Plug BDB1B or Endo to power and see Android start without USB cable.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/431/
Reviewed by: mbolivar