nuttx: max17050: Check for Power-On Reset
During initialization, check if the POR bit is set. This indicates
the device lost power and should be reconfigured. While POR is
now checked, the register init required is still not performed.

In addition, to reduce log spew, switch to dbg() function for
logging. This is controlled by the debug option in the defconfig.

Signed-off-by: Greg Meiste <w30289@motorola.com>