apps: Add pwm application to control PWM0 and PWM1 output on the bridges
One of the requirements for BDB2A bring-up is to have a means to test the
PWM0 and PWM0 engines built into the Bridge ASICS.

This commit adds a new "pwm" application to the bridge firmware, conditioned
on ARA_BRIDGE_PWM being defined.  Via "make menuconfig", that option
may be selected by enabling "Application Configuration/Ara Applications/Ara
Bridge PWM control".  The Kconfig machinery should enable ARA_BRIDGE_PWM
when DEBUG is enabled, and not enable the option when DEBUG is not enabled.
The present commit touches-up our non-endo bridge configs accordingly.

The pwm application may be invoked from a bridge's serial console, like so:

pwm <pwm-num> <secs> : output PWM for <sec> seconds
                       where  <pwm-num> is '0' for PWM0 or '1' for PWM1

The present version of the application outputs a fixed waveform of 300Khz
with a 50% duty cycle.

For PWM1 to output from a Bridge ASIC, it turns out that two bits of the
PINSHARE register must be clear: TSB_PIN_GPIO9 and TSB_PIN_UART_CTSRTS.
The pwm application records the state of these bits, clears them if need
be, and then restores their state upon exit.

The Bridge PWM hardware needs to be initialized, and the NuttX PWM driver
needs to be registered.  This commit adds an implementation of the
pwm_devinit () interface provided by various existing Nuttx architectures,
for use by the bridges on the BDB.  This interface may safely be called
multiple times, and is invoked by the pwm application to ensure that the
PWM driver is ready to use.

While debugging the pwm application, I discovered that the bridge firmware
was critically short on preallocated watchdog (wdog) timer structures that
NuttX uses for various timing purposes: the bridge configurations preallocated
a total of 6, and reserved none for allocation from atomic contexts.  I
changed CONFIG_PREALLOC_WDOGS from 6 to 8, and CONFIG_WDOG_INTRESERVE from 0
to 1 for the BDB bridge defconfigs, which matches the existing configurations
for the SVC.  This provided enough headroom to allow the pwm application to
sleep while waiting for the specified number of seconds to elapse.

This commit adds an interface to retrieve the content of the PINSHARE register,
tsb_get_pinshare(), as otherwise each application that needs to reference
PINSHARE will end up duplicating the code, potentially with a hardcoded
register address.

Testing Done: Build bdb/apb1 with the defconfig being committed and load and
    run on the image on a BDB1B with a scope attached to Pin 1 (PWM0) and Pin 3
    (PWM1) of J40.  Invoke the command "pwm 0 5" and see PWM0 output a square
    wave for 5 seconds.  Invoke the command "pwm 1 20" and see PWM1 output a
    square wave for 20 seconds.  Also verify an error-free build of bdb/apb2
    and bdb/apb1_nsh_pl2303.
Bug Number: SW-618
Review URL: https://reviews.leaflabs.com/r/442/
Reviewed by: mbolivar

Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>