bridges: Set TRACE drive strength when enabling ETM PINSHARE at Boot
On the Toshiba Bridge ASICS, the drive strength of the TRACE signals can
be configured to one of three values:
   2 milliamps
   4 milliamps (the power-on default)
   8 milliamps

The etm application allows control of instruction tracing via the serial
console, but for specialized diagnostic test builds (e.g. while debugging or
using the ETM facility), we may wish to enable ETM PINSHARE at boot via
the TSB_PINSHARE_ETM option.

Per a recommendation by Perry, the present commit causes the drive strength
of the TRACE signals to be set to the maximum value of 8ma when
TSB_PINSHARE_ETM is enabled.  The user can make a selection of 2ma or 4ma
instead, if desired, for flexibility during test.

Testing Done: Build bdb/apb1 with TSB_PINSHARE_ETM enabled and test with 8ma,
4ma, and 2ma TRACE drive strengths on AP Bridge 1 of a BDB1B.  Halt the CM3
via gdb, examine the value of the value of the IO_DRIVE_STRENGTH1 register,
and confirm that it corresponds to the configured drive strength.

Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/458/
Reviewed by: mbolivar

Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>