apps: svc: Pass interface id instead of module id during setup phase
There can be more than one interface on a module and we need to know the
interface for which the event has occurred.

But at the same time we may not need the module id at all. During
initial phase when AP is probed, the AP will receive the unique Endo id
which shall be enough to draw relationships between interface and module
ids.

Now that we have interface id, update rest of the code to use it.

Testing Done: gpio, i2c, and i2s all work
Bug Number: SW-533
Review URL: https://reviews.leaflabs.com/r/511/
Reviewed by: mgreer

Tested-by: Mark A. Greer <mgreer@animalcreek.com>
Tested-by: Marti Bolivar <mbolivar@leaflabs.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>