Add pinshare app to get, set, and clear PINSHARE register
For test and diagnostic purposes, it is useful to be able to interrogate
and modify a bridge's PINSHARE register via the serial console.  The
present commit adds a new "pinshare" application for this purpose, with the
usage:

    pinshare r|g                       : display current value of PINSHARE
    pinshare s  <hex bitmask to set>   : set specified bits of PINSHARE
    pinshare c  <hex bitmask to clear> : clear specified bits of PINSHARE

To avoid a user accidentally denying his or her self access to the serial
console, an attempt to clear the UART RX/TX Pinshare bit 0 is blocked, so
for example,
    pinshare c F
is treated as:
    pinshare c E
and the user is notified as to what happened, and why.

Testing Done: Developed and used during BDB2A bringup using Marti's
    bdb2_bringup branch.  Cherry-picked to master and built using
    the config changes in r583 for ES1 and ES2 bridges.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/582/
Reviewed by: morgang

Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>