ara: move SVC's tsb_unipro.h to <nuttx/greybus/tsb_unipro.h>
This is a header that contains Toshiba extensions to the UniPro DME
attribute space. It is needed by both the SVC and the Toshiba bridge
ASICs to manage UniPro ports on the switch and within the bridges,
respectively. Put it here so they can both find it.

Testing Done: build tested BDB2A config; grepped for occurrences of
              tsb_unipro.h and updated them appropriately.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/569/
Reviewed by: Fabien Parent, Jean Pihet