configs: ara: add new debug configuration for bridges
The current bridges configuration are just a mess, so create something a new
debug configuration in the ara configuration folder.

The debug-generic configuration is intended to be a generic image that
boots and gets you to a shell.

Testing Done: build es1 configuration and boot it
Bug Number: none
Review URL: https://reviews.leaflabs.com/r/571/
Reviewed by: fparent, mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>
Signed-off-by: Perry Hung <perry@leaflabs.com>