tsb: dwc2: fix byte ordering in dwc2 driver
The OS layer for the dwc2 driver define functions for byte ordering conversion.
This layer was using a macro __LITTLE_ENDIAN to determine if the machine
is little endian or big endian, but this macro is never defined anywhere
for the TSB machine, which means the code was assuming the machine was
big endian which is wrong.
Fix that by using the macros defined in byteorder.h instead.

Testing Done: getting the right values with the usb-host test app
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/622/
Reviewed by: ptitiano

Signed-off-by: Fabien Parent <fparent@baylibre.com>