dwc2: workaround: use address dma for hcd driver
For some unknown reasons, urb_enqueue's with DMA descriptor enabled is not
working, i.e. the transfer is configured and started but we never
receive anything on the device. Use Address DMA when using the driver in host
mode.

Note: This is worth investigating in the future why descriptor DMA is not
working.

Testing Done: enqueued something, and got an interrupt back.
Bug Number: SW-768
Review URL: https://reviews.leaflabs.com/r/806/
Reviewed by: mgreer

Signed-off-by: Fabien Parent <fparent@baylibre.com>