configs: stm32l4discovery: Add LED support
Unlike the F4 discovery, the L4 discovery only has 2 LEDs.
Because of this, the green LED was chosen as LED1 and will
be on when things are working normally. The red LED is LED2,
and will light (along with LED1) for interrupts, signals, and
assertions. If a panic, only LED2 will be lit to indicate a
fatal error (red light of death).

This addition aids in debugging.

Change-Id: I864f3db822015ddec6f8fb73189c68771bcc9677
Signed-off-by: Greg Meiste <w30289@motorola.com>