Ara: SVC: BDB1B/2A: Convert Switch to the regulator framework
Allow to ref-count the uses of power supplies and so fixes the
dependency issue of the I/O Expanders requiring 1P8 before the
switch is enabled. The switch init code requests the power supplies
regulator (which might already be enabled for the I/O Expanders) and
releases the switch reset.

Testing Done: build and run the SVC on BDB1B/2A; check power supplies
with the multimeter; check that the Switch functions correctly.
Bug Number: SW-825
Review URL: https://reviews.leaflabs.com/r/721/
Reviewed by: ptitiano

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>