arm: fix return value of semihosting_read
In the ARM semihosting specs, it is specified that read return the number
of byte not filled in the input buffer. This is the opposite of the POSIX.1
spec. Fix the return value to follow POSIX.1 spec.

Testing Done: trivial
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/752/
Reviewed by: gregkh, ptitiano

Signed-off-by: Fabien Parent <fparent@baylibre.com>