apps: ara: Don't unconditionally make some Ara apps
A few Ara applications have duplicate CNTXTDIRS entries in
the Makefile that are outside of the 'if' that builds them
only when 'CONFIG_NSH_BUILTIN_APPS' is defined.  The duplicates
that are outside the 'if' should be removed.

Testing Done: Trivial
Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/851
Reviewed by: fparent, mbolivar

Signed-off-by: Mark Greer <mgreer@animalcreek.com>