Makefile: quiet the builds a bit more
The constant notification of make entering and leaving directories is
annoying, so remove that message by adding to the MAKEFLAGS option.

Note, now we see a lot of still annoying "nothing to be done for all"
type messages on a build, but those can be fixed up later if so desired.

Testing Done: Lots of rebuilds, all succeeding
Bug Number: https://reviews.leaflabs.com/r/856/
Review URL:
Reviewed by: mbolivar, mgreer

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>