Makefile: target cleanups to make things quiet
Provide "@true" targets for makefile references that have nothing to do.
This removes the "Nothing to do be done for ..." messages when all is
build properly in the respective directory.

Now the builds almost look "quiet" for when nothing is to be done.

Testing Done: lots of builds
Bug Number:
Review URL: https://reviews.leaflabs.com/r/857/
Reviewed by: mbolivar, mgreer

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>