greybus-utils: manifest: Don't allocate memory on the heap for the manifest
There is no point in allocating memory on the heap for the manifest since
the manifest is already in memory and we don't intend to modify it.

Remove the logic to copy the data into dynamically allocated memory.

Testing Done: Reproduced gpio demo
Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/932/
Reviewed by: mark

Signed-off-by: Benoit Cousson <bcousson@baylibre.com>