arm: ara: define byte order api outside of armv7m folder
Several drivers have implicit dependencies on the arm v7m architecture
because they are using a file defined with a patch that includes the
architecture name.

Move the byteorder.h in a folder that doesn't require add the name of
the architecture in the include path.

Testing Done: trivial
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/944/
Reviewed by: bcousson, mgreer

Signed-off-by: Fabien Parent <fparent@baylibre.com>