ara: nuke bdb, endo configs directories from orbit
It's the only way to be sure.

The code in nuttx/configs/{bdb,endo} is stale and increasingly hard to
maintain. We have done what we need to get rid of them now, leaving
only configs/ara, The One True Configs Directory (TOTCD):

- It's safe to delete nuttx/configs/bdb now that the wiki has been
  updated to reflect the new ES2 configs that non-firmware developers
  will need to smoke-test their boards.

- It's safe to delete nuttx/configs/endo as the final Google I/O demo
  code is kept safely in an out of tree branch.

- The only old config we still want is bdb/apb1_nsh_pl2303, because
  that's how we do APBridgeA USB device bringup testing on new boards.

  Move that to nuttx/configs/ara/bridge, intentionally dropping its
  unused and cargo-culted setenv.sh.

  This requires minimal defconfig hand-editing to get it to build
  again (make menuconfig failed because its CONFIG_ARCH_BOARD setting
  pointed at a nonexistent directory).

Also do a minimal update of build_ara_image.sh to keep that
working. Now that everything is under "nuttx/configs/ara", a next step
would be to make it aware of that fact. But that might break people's
scripts, so let's not do it yet.

configs/ara/{apb1,apb2} still contains "demo" code for an ES1 DSI
tunnel. It seems harmless to keep that around as long as it builds;
we'll want it later for porting to future hardware.

Testing Done: build_ara_image.sh -j 6 all
Bug Number: SW-535, SW-838
Review URL: https://reviews.leaflabs.com/r/984/
Reviewed by: Darryl Newberry, Mark Greer, Fabien Parent

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>