kconfig: Change 'CONFIG_APBRIDGE' to 'CONFIG_APBRIDGEA'
The config variable/macro, 'CONFIG_APBRIDGE', is used to build
code that runs on APBridges on AP Modules only.  These bridges
are called APBridgeAs.  CONFIG_APBRIDGE is not for APBridges that
are not on AP Modules.  Those are referred to as APBridgeEs.

So, CONFIG_APBRIDGE is for APBRidgeAs only and CONFIG_GPBRIDGE is
for all of the other bridges (i.e., all APBridgeEs and GPBridges).
This is confusing so change CONFIG_APBRIDGE to CONFIG_APBRIDGEA to
indicate that it is for APBridgeAs only.

Also add Kconfig help information to help users when deciding
whether to enable CONFIG_APBRIDGEA.

Testing Done: Ran build_ara_image.sh
Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/1146/
Reviewed by: greg-kh

Signed-off-by: Mark Greer <mgreer@animalcreek.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>