svc: wait for and handle ap module boot
Register the SVC for mailbox events and wait for the AP module to boot.
When the AP enumerates APBA, it will write into the switch-side mailbox,
notifying the SVC that it is ready, as well as it's position in the
endo.

HACK: Hotplug events are faked atm.

Testing Done:
Bug Number: SW-319
Review URL: https://reviews.leaflabs.com/r/1204
Reviewed by: Greg Kroah-Hartman

Signed-off-by: Perry Hung <perry@leaflabs.com>
Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>