tsb_unipro_es1: fix up DBG_UNIPRO() to catch printf errors
DBG_UNIPRO() was not checking for printf errors, especially when
UNIPRO_DEBUG is not enabled.  When this was changed, a few errors were
found in DBG_UNIPRO() usages, so fix those up as well so as to not
pollute the build with warnings.

Testing Done: lots of build tests
Review URL: https://reviews.leaflabs.com/r/1211/
Reviewed by: mbolivar, fparent

Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>