nsh: notify the user when task limit is exceeded
When task limit is exceeded and the user wants to run a built-in command,
nsh fails and emits the following message:

    nsh> gpio
    nsh: gpio: command not found

This is misleading as the real reason is the exhaustion of available PIDs.

Make nsh output a clearer message in the above case.

Testing Done: Proper message is displayed.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1214/
Reviewed by: Fabien Parent, Alexandre Bailon

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>