gb_loopback: rework gb_loopback's front-end and API
This patch introduces several major changes in gb_loopback's internal and
external interface. The general idea is to separate the logic between the
driver and gbl application. We introduce two separate structures for
loopbacks: one internal for the driver with which we associate the cport
upon registration and a second one in the application that we use to
supervise the commands.

We also introduce a new command-line interface. The user now has to specify
whether he wants to start a command, stop it or view the statistic data
gathered by the service.

Changes in the driver API:
- removed most fields from struct gb_loopback in order to further reduce
  the mixing of logic between the application and the driver,
- implemented gb_loopback_get_cports() function which allows gb_loopback
  to retrieve the information about the cports on which the loopback driver
  is registered,
- improved locking,
- made struct gb_loopback private in order to hide the driver's logic from
  the application,
- inlined the header with new loopback API,

Testing Done: Performed extensive testing on bdb2a: tested all arguments
              from APB2 and observed the traffic on APB1. Verified the
              output of gbl status.
Bug Number: SW-957
Review URL: https://reviews.leaflabs.com/r/1276/
Reviewed by: Marti Bolivar

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>