tsb: uart: Fix UART rx and tx interrupt handlers
The rx interrupt handler only received one byte at a time
if the character timeout bit was set (UA_INTERRUPT_ID_TO).
Change the handler to empty the rx fifo as long as their
is room in the destination buffer.

The tx interrupt handler always sent a full fifo of data
even if the source buffer was less than the fifo size.
Change the handler to return if all of the data from the
source buffer is sent.

Signed-off-by: Mike Corrigan <corrigan@motorola.com>