Ara: SVC: vreg: Allow the declaration of empty vregs list
A vreg structure can be declared as empty, i.e. without any regulator
control pins. This happens in the case an SVC interface has been
declared with a NULL value in vreg_data because it has no vreg
associated to it.

In that case the vregs operations are no-ops and are not failing.

Testing Done: Build for all SVC configs; Test on BDB2A with empty
vreg(s)
Bug Number: SW-82
Review URL: https://reviews.leaflabs.com/r/1345/
Reviewed by: mgreer, mbolivar, fparent

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>