ara: usbdev: correctly free buffer request on errors
When an error happens during the process of a usb control setup packet, the
request buffer allocated is not always freed correctly, i.e. inserted in the
list of free buffers.

This fix a hardfault that happens when receiving request that the firmware is
not able to answer.

Testing Done: checked that no more hardfault happens when receiving several
              packet that the FW can't respond to.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1374/
Reviewed by: abailon, mgreer

Signed-off-by: Fabien Parent <fparent@baylibre.com>