build_ara_image.sh: make script usable as a library
In order to use this script for out of tree, make it usable as a library, i.e.
only execute the script only if the file has been executed and not sourced.

Testing Done: used the script to build images.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1380/
Reviewed by: newberry_darryl, mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>