device: Remove union of device type operations
To help facilitate out-of-tree builds and make the
device.h file cleaner, remove the union of device
type operations and replace with a void pointer.
Update the macros that use the union.

This change forces the structure defining device
type operations to be of the form:

    struct device_<type>_type_ops {
        ...
    };

but all of the device type operations defined so
far already conform to that format and it doesn't
seem unreasonable to force it upon any new device
types.

Testing Done: Ran Ran build_ara_image.sh, ran GPIO test on a bdb2b
Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/1407/
Reviewed by: fparent, mbolivar, mgirling

Signed-off-by: Mark Greer <mgreer@animalcreek.com>