mods: add unipro_cport_count()
greybus core uses unipro_cport_count to determine the number of
cports available.  Outside of unipro, the definition is arbitrary.
Add unipro_cport_count to network.c where we were using the
arbitrary macro in the first place.  This function is not used
by unipro implementations and should be safe.

Signed-off-by: Jim Wylder <jwylder@motorola.com>