tsb: uart: fix possible memory corruptions
The UART driver is not checking whether the RX or TX path has been enabled.
We can get interrupts for TX or RX even when they are not enabled, when
this happen some memory corruption happens because we are using a buffer address
that has not been initialized to a valid address.

Testing Done: generated a condition that would make the memory corruption
              occur, noticed an hardfault, applied this patch, noticed
              that the hardfault is gone.
              Checked that the uart demo works.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/1495/
Reviewed by: mbolivar

Signed-off-by: Fabien Parent <fparent@baylibre.com>