gpbridge init: distinguish between booting from flash or bootrom
The nuttx firmware needs to distinguish between being loaded from SPIROM or via
the bootrom/Greybus Firmware Protocol.  In the former case, the gpbridge app
should poke the switch mailbox once machine initialization has finished to say
it's ready for Greybus connections, causing the SVC to generate a hotplug event.
In the latter case, the bootrom did this for us, and a new hotplug event will
be wrong.

Set and read a TSB-specific static variable denoting when nuttx is functioning
as stage 1 or stage 2.

Testing Done: ara-bridge-es2-debug-generic on apb2 with *and* without bootrom
Bug Number:
Review URL: https://reviews.leaflabs.com/r/1501/
Reviewed by: Jun Li <li_jun@projectara.com>

Signed-off-by: Eli Sennesh <esennesh@leaflabs.com>