DWC USB OTG: let the DMA continue on BNA error
If the SG DMA open a descriptor with a status different from
HOST_READY, then the DMA generate a BNA error.
Once the CPU managed the cause of BNA, the default behavior
of DMA is to restart with the first descriptor.
That is breaking the request management that is expecting circular
operations.
Let the DMA restart with the descriptor that cause the BNA.

Testing Done:
- Use loopback-perf tool to mesaure performances with and without the serie:
  no performances changes
- Use usb_test to stress test: no issue except sometime there are out of requests on
  UniPro rx path (not related to this serie).
Bug Number: sw-1006
Review URL: https://reviews.leaflabs.com/r/1569/
Reviewed by: fparent

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>