DWC USB OTG: Update BNA handleing
Currently, if the SG DMA generate a BNA,
the CPU reset all descriptors.
This is not a correct way to handle error because of the ring descriptors.
Usualy, BNA is generated because the DMA has used all descriptor and
the CPU didn't get time to complete the transfer and push pack the request
to the list (and make it available to DMA).
The way to handle the BNA is to disable the endpoint until a new request is available.

Testing Done:
- Use loopback-perf tool to mesaure performances with and without the serie:
  no performances changes
- Use usb_test to stress test: no issue except sometime there are out of requests on
  UniPro rx path (not related to this serie).
Bug Number: sw-1006
Review URL: https://reviews.leaflabs.com/r/1571/
Reviewed by: fparent

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>