Pull up GPIO0 on APBridgeA so that Helium will boot on SDB
On the SDB, the EXP_nRST_BTN on the 96Boards Expansion Header is wired to GPIO0
of APBridgeA, and this signal was floating, leading to boot failure on SDB
S/N's 42 and 10, but not on S/N 3.

Rob Herring indicates that EXP_nRST_BTN must be pulled up or driven high.
This patch programs APBridgeA to pull up GPIO0, which is a less aggressive
approach than actively driving the GPIO, and avoids the need to enable the GPIO
subsystem early in the NuttX boot sequence.

APBridgeA GPIO0 is wired to a debug header on BDB2{A,B}, and has no other
function on those boards, so I don't anticipate a regression as a result of
this patch.

Testing Done: Bring up nuttx master on SDB S/N 042 with Helium attached to the
    96Boards Expansion Header, verify that Helium fails to boot.  Apply patch,
    rebuild and reload APBridgeA, and verify that Helium boots.
Bug Number: Trivial
Review URL: https://reviews.leaflabs.com/r/1529/
Reviewed by: mbolivar, esennesh
Signed-off-by: Mitchell Tasman <tasman@leaflabs.com>