Ara: SVC: Remove dead code for Unipro devices detection
Remove unused code. At boot time before the interfaces are up the link
status is always 0 and so no devices are detected.

The devices detection is to be done using the Wake&Detect lines and the
hotplug mechanism with the AP

Note: the link status can be queried anytime with 'svc s' on the console.

Testing Done: Boot on SDB
Bug Number: SW-923
Review URL: https://reviews.leaflabs.com/r/1577/
Reviewed by: mbolivar

Signed-off-by: Jean Pihet <jean.pihet@newoldbits.com>