greybus: control: add reboot bootloader for flashing
Add a control protocol message GB_CONTROL_TYPE_REBOOT_FLASH.
When this request is received, the state should be set such
that, on the next boot, the bootloader will remain in flash
mode.  This command does this by setting a predetermined bit
pattern in the last page of flash memory.  It does not
instantiate the reboot.

Change-Id: I67d931e834bf27b3a6d0fd268d307fa1a8b87d75
Signed-off-by: Jim Wylder <jwylder@motorola.com>