power: bq24292: Switch from pthread to work queue to save RAM
Using pthreads is expensive. Each pthread creates a new stack in
RAM, and uses up a task slot, which is artifically limited in Nuttx.
Increasing this limit uses more RAM, so we would like to avoid that.

It's likely that this change removes the dependency on user space
for driver initialization, but leaving initialization as-is for now.

Change-Id: If2ccea349868c30bf0d7742a41f1487359f57972
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/786116
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Dmitriy Gruzman <dgruzman@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>