usb_test: add timeout argument
Currently, the timeout of a transfer is hardcoded to 1s.
But if we are performing a lot of transfers, 1s may be too short.
Add an argument to configure the timeout.

Testing Done: usb_test can change the USB timeout
Bug Number: SW-1007
Review URL: https://reviews.leaflabs.com/r/1755/
Reviewed by: fparent

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>