loopback: move loopback definitions to shared header
In order to facilitate stuffing of timestamping data in the APBridge it is
necessary to make the APBridge aware of the location of where to stuff data
on designated CPorts. To facilitate with one header owning the definition
of the loopback response header, this patch moves the definition of the
loopback data structures from nuttx/drivers/greybus/loopback-gb.h to
nuttx/include/nuttx/greybus/loopback.h - which the APBridge and GPBridge
firmware images can both then include.

Testing Done: Tested with in-tandem kernel header change.
Bug Number: SW-1097
Review URL: https://reviews.leaflabs.com/r/1767
Reviewed by: Eli Sennesh, Mitch Tasman
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>